\name{GSE142512-package}
\alias{GSE142512-package}
\docType{package}
\title{GSE142512 methylation data in ExperimentHub}
\description{
The \pkg{GSE142512} package provides ExperimentHub access to processed DNA
methylation resources derived from GEO accession GSE142512.
}
\details{
The hosted resources are \code{RangedSummarizedExperiment} objects created
from Illumina HumanMethylation450 and Infinium MethylationEPIC
\code{GenomicRatioSet} objects. They contain beta, M, and CN assays with hg19
genomic ranges and sample-level phenotype metadata.

The resources are hosted on Zenodo at
\url{https://zenodo.org/records/21198533} and are registered for access through
\pkg{ExperimentHub}.

See \code{browseVignettes("GSE142512")} for usage examples.
}
\seealso{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE142512}
}
\keyword{datasets}
